+++ pkg/applicationserver/applicationserver.go
@@ -131,6 +131,10 @@ func New(c *component.Component, conf *Config) (as *ApplicationServer, err error
 		interopID:        conf.Interop.ID,
 		endDeviceFetcher: conf.EndDeviceFetcher.Fetcher,
 	}
+
+	if as.endDeviceFetcher == nil {
+		as.endDeviceFetcher = &NoopEndDeviceFetcher{}
+	}
 	retryIO := io.NewRetryServer(as)
 
 	as.grpc.asDevices = asEndDeviceRegistryServer{
@@ -935,6 +939,13 @@ func (as *ApplicationServer) handleUplink(ctx context.Context, ids ttnpb.EndDevi
 		uplink.LastAFCntDown = dev.Session.LastAFCntDown
 	}
 
+	isDev, err := as.endDeviceFetcher.Get(ctx, ids, "locations")
+	if err != nil {
+		logger.WithError(err).Warn("Failed to retrieve end device locations")
+	} else {
+		uplink.Locations = isDev.GetLocations()
+	}
+
 	// TODO: Run uplink messages through location solvers async (https://github.com/TheThingsNetwork/lorawan-stack/issues/37)
 	return nil
 }
@@ -950,6 +961,14 @@ func (as *ApplicationServer) handleSimulatedUplink(ctx context.Context, ids ttnp
 	if err != nil {
 		return err
 	}
+
+	isDev, err := as.endDeviceFetcher.Get(ctx, ids, "locations")
+	if err != nil {
+		log.FromContext(ctx).WithError(err).Warn("Failed to retrieve end device locations")
+	} else {
+		uplink.Locations = isDev.GetLocations()
+	}
+
 	return as.decodeUplink(ctx, dev, uplink, link.DefaultFormatters)
 }